pyproject.toml
[project]
name = "domain-watchdog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-whois>=0.9.5",
    "requests>=2.32.3",
]
